+++ src/ray.cpp
@@ -7,7 +7,7 @@ ray::ray()
 {}
 ray::ray(const point3& origin, const vec3& direction) : origin_(origin), direction_(direction)
 {}
-point3 ray::at(double t) const {
+point3 ray::at(float t) const {
     return origin_ + t*direction_;
 }
